Move shared api to the startup/both folder
