--- imports/startup/api.js
@@ -1,1 +1,0 @@
-import '../api/Documents/methods';
