+++ imports/startup/both/api.js
@@ -1,0 +1,1 @@
+import '../../api/Documents/methods';
