+++ imports/startup/client/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Meteor } from 'meteor/meteor';
 import App from '../../ui/layouts/App/App';
-import '../api';
+import '../both/api';
 import './api';
 
 import '../../ui/stylesheets/app.scss';
